+++ Light Engine/src/Engine/Events/WindowEvents.h
@@ -11,14 +11,14 @@ namespace Light {
 	class WindowResizedEvent : public Event
 	{
 	private:
-		uint16_t m_Width, m_Height;
+		int m_Width, m_Height;
 	public:
 		WindowResizedEvent(uint16_t  width, uint16_t  height): m_Width(width), m_Height(height) {}
 
-		inline uint16_t GetWidth () const { return m_Width;  }
-		inline uint16_t GetHeight() const { return m_Height; }
+		inline int GetWidth () const { return m_Width;  }
+		inline int GetHeight() const { return m_Height; }
 
-		inline std::pair<uint16_t, uint16_t> GetSize() const { return { m_Width, m_Height }; }
+		inline std::pair<int, int> GetSize() const { return { m_Width, m_Height }; }
 
 		std::string GetLogInfo() const override
 		{
@@ -34,14 +34,14 @@ namespace Light {
 	class WindowMovedEvent : public Event
 	{
 	private:
-		int16_t m_PositionX, m_PositionY;
+		int m_PositionX, m_PositionY;
 	public:
-		WindowMovedEvent(int16_t xpos, int16_t ypos): m_PositionX(xpos), m_PositionY(ypos) {}
+		WindowMovedEvent(int xpos, int ypos): m_PositionX(xpos), m_PositionY(ypos) {}
 
-		inline uint16_t GetX() const { return m_PositionX; }
-		inline uint16_t GetY() const { return m_PositionY; }
+		inline int GetX() const { return m_PositionX; }
+		inline int GetY() const { return m_PositionY; }
 
-		inline std::pair<uint16_t, uint16_t> GetPosition() const { return { m_PositionX, m_PositionY }; }
+		inline std::pair<int, int> GetPosition() const { return { m_PositionX, m_PositionY }; }
 
 		std::string GetLogInfo() const override
 		{ 
@@ -72,7 +72,6 @@ namespace Light {
 		EVENT_CATEGORY(EventCategory_Window)
 	};
 
-	// (win32): Also gets called with window resize event on restored state
 	class WindowRestoredEvent : public Event
 	{
 	public:
@@ -91,7 +90,7 @@ namespace Light {
 		EVENT_CATEGORY(EventCategory_Window)
 	};
 
-	class WindowMinimizedEvent : public Event // minimized
+	class WindowMinimizedEvent : public Event
 	{
 	public:
 		std::string GetLogInfo() const override { return "Window minimized"; }
